backend/pyproject.toml
[project]
name = "memoryos"
version = "0.1.0"
description = "Enterprise-grade, MCP-native self-improving memory platform for AI agents."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "email-validator>=2.2.0",
  "httpx>=0.27.0",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.1",
  "redis>=5.0.8",
  "prometheus-client>=0.20.0",
  "sentence-transformers>=3.0.1",
  "huggingface-hub>=0.24.6",
  "transformers @ git+https://github.com/huggingface/transformers@v4.56.0-Embedding-Gemma-preview",
  "pgvector>=0.3.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
